JSSC-74/JSSC-75/JSSC-76 그룹인트로 리팩토링
