+++ src/Component/CTA/CTAContainer.tsx
@@ -1,19 +1,24 @@
 import axios from 'axios';
-import { useContext, useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { useLocation, useNavigate, useParams } from 'react-router-dom';
 import { styled } from 'styled-components';
-import { MyListContext, MyListContextType } from '../../API/Context';
 import { postJoin } from '../../API/joinedMinds';
 import { getMyList } from '../../API/Mind';
 import { EXPIRED_TOKEN, INVALID_TOKEN } from '../../constant/error';
-// TODO: api: 참여하기 요청보내기 /joined-minds/{joined_mind_id}
-/** 2023-08-22 CTAContainer.tsx - 참여하기 버튼 */
+
+const CTAContainer = (): JSX.Element => {
+  return (
+    <CTAContainerS>
+      {/* <GNB /> */}
+      <JoinButtonCTA />
+    </CTAContainerS>
+  );
+};
+
 const JoinButtonCTA = (): JSX.Element => {
   const navigate = useNavigate();
   const { mindId } = useParams();
-  const [isLogin, setIsLogin] = useState(false);
   const [validJoin, setValidJoin] = useState('true');
-  const { myList, setMylist } = useContext<MyListContextType>(MyListContext);
   const location = useLocation();
 
   useEffect(() => {
@@ -55,12 +60,6 @@ const JoinButtonCTA = (): JSX.Element => {
   );
 };
 
-/** 2023-08-22 CTAContainer.tsx - 작심 인증하기 버튼 */
-const PostButton = (): JSX.Element => {
-  const navigate = useNavigate();
-  return <MissionButtonS onClick={() => navigate(`/uploadPost/1`)}>작심 인증하기</MissionButtonS>;
-};
-
 /** 2023-08-28 CTAContainer.tsx - 작심 인증하기 버튼 */
 const BackButton = (): JSX.Element => {
   const navigate = useNavigate();
@@ -70,16 +69,6 @@ const BackButton = (): JSX.Element => {
 /** 2023-08-22 CTAContainer.tsx - 인증하기 버튼 */
 const SubmitButtonCTA = (): JSX.Element => {
   return <CTAButtonS valid={'true'}>인증하기</CTAButtonS>;
-};
-
-/** 2023-08-22 CTAContainer.tsx - CTA 참여하기 + GNB */
-const CTAContainer = (): JSX.Element => {
-  return (
-    <CTAContainerS>
-      {/* <GNB /> */}
-      <JoinButtonCTA />
-    </CTAContainerS>
-  );
 };
 
 /** 2023-08-22 CTAContainer.tsx - CTA 참여하기 + GNB */
@@ -95,59 +84,54 @@ const ErrorCTA = (): JSX.Element => {
   );
 };
 
-export { JoinButtonCTA, PostButton, SubmitButtonCTA, CTAContainer, BackButton, ErrorCTA };
+export { JoinButtonCTA, SubmitButtonCTA, CTAContainer, BackButton, ErrorCTA };
 
 /** 2023-08-22 CTAContainer.tsx - CTA 참여하기 + GNB */
 const CTAContainerS = styled.div`
   position: sticky;
-  bottom: 0;
   display: flex;
   flex-direction: column-reverse;
-  gap: 1rem;
+`;
+
+const ErrorCTAS = styled.div`
+  position: absolute;
+  bottom: 0rem;
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  max-width: 500px;
 `;
 
 /** 2023-08-22 CTAContainer.tsx - 버튼 공통 스타일 */
 const LinkButtonS = styled.button`
-  height: 3.5rem;
-  border-radius: 1.88rem;
-  margin-top: 1.25rem;
   display: flex;
   justify-content: center;
   align-items: center;
-`;
-
-const ErrorCTAS = styled.div`
-  position: fixed;
-  bottom: 1rem;
-  width: 100%;
-  display: flex;
-  flex-direction: column;
+  height: 3.5rem;
+  border-radius: 1.88rem;
+  margin-top: 1.25rem;
 `;
 
 /** 2023-08-22 CTAContainer.tsx - CTA 버튼(참여하기, 인증하기) */
 const CTAButtonS = styled(LinkButtonS)<{ valid: string }>`
+  position: sticky;
+  margin: 1rem;
   background-color: ${(props) =>
     props.valid === 'true' ? 'var(--color-main)' : 'var(--color-disabled2)'};
-  margin: 1rem;
-  margin-bottom: 1rem;
-  position: sticky;
-  bottom: 0rem;
-  font-size: var(--button-mid);
-  color: ${(props) => (props.valid === 'true' ? 'var(--font-color1)' : 'var(--color-disabled1)')};
   font-size: 1rem;
+  color: ${(props) => (props.valid === 'true' ? 'var(--font-color1)' : 'var(--color-disabled1)')};
 `;
 
 /** 2023-08-22 CTAContainer.tsx - 작심 인증하기 버튼 */
 const MissionButtonS = styled(LinkButtonS)`
+  width: 100%;
   border: 0.1rem solid;
   border-color: var(--color-main);
   font-size: var(--button-mid);
-  width: 100%;
 `;
 
 /** 2023-08-28 CTAContainer.tsx - 이전 페이지로 버튼 */
 const BackButtonS = styled(MissionButtonS)`
   height: 2.5rem;
   width: 11.25rem;
-  font-size: var(--button-mid);
 `;
