--- src/Component/Mission/DivideBa.tsx
@@ -1,11 +1,0 @@
-import { styled } from 'styled-components';
-
-const DivideBaS = styled.div`
-  width: 100vw;
-  margin: 1.25rem 0;
-  height: var(--height-gap);
-  background-color: var(--color-line);
-  width: 100%;
-`;
-
-export default DivideBaS;
