+++ src/Component/Mission/MissionTab.tsx
@@ -103,6 +103,6 @@ const MissionSingleWideS = styled.div`
   border: 1px solid;
   width: fit-content;
   p {
-    font-size: 11px;
+    font-size: var(--button-small);
   }
 `;
