+++ src/Page/GroupPage/GroupPageBarrel.ts
@@ -5,13 +5,10 @@ import { useNavigate } from 'react-router-dom';
 import useLoginCheck from '../../Hooks/useLoginCheck';
 import { GroupHeader } from '../../Component/Mission/GroupHeader';
 import Comment from './Comment';
-import DivideBaS from '../../Component/Mission/DivideBa';
+import { DivideBaS } from '../../Component/Mission/GroupArticle';
 import GroupPost from './GroupPost';
-// import Mind from './Mind';
-import GroupArticle from '../../Component/Mission/GroupArticle';
 
 export {
-  // Mind
   useState,
   useEffect,
   styled,
@@ -21,5 +18,4 @@ export {
   Comment,
   DivideBaS,
   GroupPost,
-  GroupArticle,
 };
